src/requirements.txt
flask==2.3.3
werkzeug==2.3.7
sqlalchemy==1.4.23
mysqlclient==2.0.3
python-dotenv==0.19.0
tensorflow==2.13.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
prometheus-client==0.11.0
gunicorn==20.1.0
sqlparse==0.4.2
python-nmap==0.7.1
scapy==2.5.0
psutil==5.9.0
requests==2.31.0
python-whois==0.8.0
paramiko==3.3.1
colorama==0.4.6
pycryptodome==3.19.0
python-magic==0.4.27
yara-python==4.3.1
